Versão final 2.0.0

+++ src/main/java/br/com/processo_seletivo/processo_seletivo/servicos/ArquivoService.java
@@ -3,6 +3,8 @@ package br.com.processo_seletivo.processo_seletivo.servicos;
 import java.util.List;
 import java.util.Optional;
 
+import javax.persistence.EntityNotFoundException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,30 +14,34 @@ import br.com.processo_seletivo.processo_seletivo.repositorios.ArquivoRepository
 @Service
 public class ArquivoService implements CrudService<Arquivo, Long> {
 
-    @Autowired
-    private ArquivoRepository arquivoRepository;
+	@Autowired
+	private ArquivoRepository arquivoRepository;
 
-    @Override
-    public Arquivo save(Arquivo entity) {
-        return arquivoRepository.save(entity);
-    }
+	@Override
+	public Arquivo save(Arquivo entity) {
+		return arquivoRepository.save(entity);
+	}
 
-    @Override
-    public Optional<Arquivo> findById(Long id) {
-        return arquivoRepository.findById(id);
-    }
-    
-    @Override
-    public List<Arquivo> findAll() {
-        return arquivoRepository.findAll();
-    }
+	@Override
+	public Optional<Arquivo> findById(Long id) {
+		return arquivoRepository.findById(id);
+	}
 
-    @Override
+	@Override
+	public List<Arquivo> findAll() {
+		return arquivoRepository.findAll();
+	}
+
     public void deleteById(Long id) {
-        arquivoRepository.deleteById(id);
+        if (arquivoRepository.existsById(id)) {
+            arquivoRepository.deleteById(id);
+            System.out.println("Arquivo com ID " + id + " deletado com sucesso");
+        } else {
+            throw new EntityNotFoundException("Arquivo com ID " + id + " não encontrado");
+        }
     }
 
-    public List<Arquivo> findByDiretorioId(Long diretorioId) {
-        return arquivoRepository.findByDiretorioId(diretorioId);
-    }
+	public List<Arquivo> findByDiretorioId(Long diretorioId) {
+		return arquivoRepository.findByDiretorioId(diretorioId);
+	}
 }
